requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg[binary]
jinja2
python-multipart
passlib[bcrypt]
python-jose[cryptography]
email-validator
redis
rq
python-dotenv
pydantic-settings
